refactor: 예약 대기 기능을 별도의 컨트롤러/서비스로 분리
