+++ src/test/java/roomescape/feature/reservation/controller/ReservationControllerTest.java
@@ -409,147 +409,4 @@ class ReservationControllerTest {
                 .andExpect(status().is4xxClientError());
         }
     }
-
-    @Nested
-    class 예약_대기_생성 {
-
-        @Test
-        void 예약_대기를_생성한다() throws Exception {
-            when(reservationMapper.toCreateCommand(any())).thenReturn(sampleCreateCommand());
-            when(reservationService.saveWaitingReservation(any())).thenReturn(sampleCreateResponse());
-
-            mockMvc.perform(post("/api/reservations/waitings")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content("""
-                        {
-                          "name": "%s",
-                          "date": "%s",
-                          "timeId": 1,
-                          "themeId": 1
-                        }
-                        """.formatted(
-                            ReservationFixture.FUTURE.getName(),
-                            ReservationFixture.FUTURE.getDate()
-                        )))
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.name", equalTo(ReservationFixture.FUTURE.getName())));
-        }
-
-        @Test
-        void 이미_대기_중인_예약이면_4xx를_반환한다() throws Exception {
-            when(reservationMapper.toCreateCommand(any())).thenReturn(sampleCreateCommand());
-            when(reservationService.saveWaitingReservation(any()))
-                .thenThrow(new GeneralException(ReservationErrorType.ALREADY_WAITING));
-
-            mockMvc.perform(post("/api/reservations/waitings")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content("""
-                        {
-                          "name": "%s",
-                          "date": "%s",
-                          "timeId": 1,
-                          "themeId": 1
-                        }
-                        """.formatted(
-                            ReservationFixture.FUTURE.getName(),
-                            ReservationFixture.FUTURE.getDate()
-                        )))
-                .andExpect(status().is4xxClientError());
-        }
-
-        @Test
-        void 예약자명이_없으면_4xx를_반환한다() throws Exception {
-            mockMvc.perform(post("/api/reservations/waitings")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(reservationRequestBody(OMITTED, VALID_DATE, "1", "1")))
-                .andExpect(status().is4xxClientError());
-        }
-
-        @Test
-        void 예약_날짜가_없으면_4xx를_반환한다() throws Exception {
-            mockMvc.perform(post("/api/reservations/waitings")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(reservationRequestBody(VALID_NAME, OMITTED, "1", "1")))
-                .andExpect(status().is4xxClientError());
-        }
-
-        @Test
-        void timeId가_없으면_4xx를_반환한다() throws Exception {
-            mockMvc.perform(post("/api/reservations/waitings")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(reservationRequestBody(VALID_NAME, VALID_DATE, OMITTED, "1")))
-                .andExpect(status().is4xxClientError());
-        }
-
-        @Test
-        void themeId가_없으면_4xx를_반환한다() throws Exception {
-            mockMvc.perform(post("/api/reservations/waitings")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(reservationRequestBody(VALID_NAME, VALID_DATE, "1", OMITTED)))
-                .andExpect(status().is4xxClientError());
-        }
-
-        @ParameterizedTest
-        @ValueSource(longs = {0L, -1L})
-        void timeId가_양수가_아니면_4xx를_반환한다(long timeId) throws Exception {
-            mockMvc.perform(post("/api/reservations/waitings")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(reservationRequestBody(VALID_NAME, VALID_DATE, String.valueOf(timeId), "1")))
-                .andExpect(status().is4xxClientError());
-        }
-
-        @ParameterizedTest
-        @ValueSource(longs = {0L, -1L})
-        void themeId가_양수가_아니면_4xx를_반환한다(long themeId) throws Exception {
-            mockMvc.perform(post("/api/reservations/waitings")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(reservationRequestBody(VALID_NAME, VALID_DATE, "1", String.valueOf(themeId))))
-                .andExpect(status().is4xxClientError());
-        }
-    }
-
-    @Nested
-    class 예약_대기_취소 {
-
-        @Test
-        void 예약_대기를_취소한다() throws Exception {
-            when(reservationService.cancelWaitingReservation(eq(1L), eq(new ReserverName(ReservationFixture.FUTURE.getName()))))
-                .thenReturn(sampleCancelResponse());
-
-            mockMvc.perform(patch("/api/reservations/1/waitings/cancel")
-                    .queryParam("name", ReservationFixture.FUTURE.getName()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id", equalTo(1)));
-        }
-
-        @Test
-        void 존재하지_않는_예약_ID이면_4xx를_반환한다() throws Exception {
-            when(reservationService.cancelWaitingReservation(eq(999L), eq(new ReserverName(ReservationFixture.FUTURE.getName()))))
-                .thenThrow(new GeneralException(ReservationErrorType.RESERVATION_NOT_FOUND));
-
-            mockMvc.perform(patch("/api/reservations/999/waitings/cancel")
-                    .queryParam("name", ReservationFixture.FUTURE.getName()))
-                .andExpect(status().is4xxClientError());
-        }
-
-        @Test
-        void 예약자명이_일치하지_않으면_4xx를_반환한다() throws Exception {
-            when(reservationService.cancelWaitingReservation(eq(1L), eq(new ReserverName("다른예약자"))))
-                .thenThrow(new GeneralException(ReservationErrorType.RESERVATION_CANCEL_FORBIDDEN));
-
-            mockMvc.perform(patch("/api/reservations/1/waitings/cancel")
-                    .queryParam("name", "다른예약자"))
-                .andExpect(status().is4xxClientError());
-        }
-
-        @Test
-        void 대기_예약이_아닌_일반_예약에_대기_취소를_시도하면_4xx를_반환한다() throws Exception {
-            when(reservationService.cancelWaitingReservation(eq(1L), eq(new ReserverName(ReservationFixture.FUTURE.getName()))))
-                .thenThrow(new GeneralException(ReservationErrorType.NOT_WAITING_RESERVATION));
-
-            mockMvc.perform(patch("/api/reservations/1/waitings/cancel")
-                    .queryParam("name", ReservationFixture.FUTURE.getName()))
-                .andExpect(status().is4xxClientError());
-        }
-    }
 }
