+++ src/test/java/roomescape/feature/reservation/service/ReservationServiceTest.java
@@ -52,7 +52,7 @@ class ReservationServiceTest {
     @BeforeEach
     void setUp() {
         ReservationMapper mapper = new ReservationMapper(new TimeMapper(), new ThemeMapper());
-        reservationService = new ReservationService(
+        reservationService = new ReservationServiceImpl(
             reservationRepository, timeRepository, themeRepository, mapper);
     }
 
@@ -261,74 +261,6 @@ class ReservationServiceTest {
             assertThatThrownBy(() -> reservationService.saveReservation(command))
                 .isInstanceOf(GeneralException.class)
                 .hasMessage("이미 예약된 날짜, 시간, 테마입니다.");
-        }
-    }
-
-    @Nested
-    class 예약_대기_생성 {
-
-        @Test
-        void 예약_대기를_생성한다() {
-            // given
-            Time time = timeWithId(1L);
-            Theme theme = themeWithId(1L);
-            LocalDate date = LocalDate.now().plusYears(1);
-            ReservationCreateCommand command = new ReservationCreateCommand(
-                new ReserverName("예약자"), date, 1L, 1L);
-            Reservation saved = Reservation.reconstruct(
-                1L, new ReserverName("예약자"), date, time, theme, ReservationStatus.WAITING);
-            when(timeRepository.findTimeByIdAndNotDeleted(1L)).thenReturn(Optional.of(time));
-            when(themeRepository.findThemeByIdAndNotDeleted(1L)).thenReturn(Optional.of(theme));
-            when(reservationRepository.existsReservationAndStatus(any(Reservation.class), any()))
-                .thenReturn(false);
-            when(reservationRepository.save(any(Reservation.class))).thenReturn(saved);
-
-            // when
-            ReservationCreateResponseDto result = reservationService.saveWaitingReservation(command);
-
-            // then
-            assertThat(result.id()).isEqualTo(1L);
-            assertThat(result.name()).isEqualTo("예약자");
-        }
-
-        @Test
-        void 같은_이름_날짜_시간_테마로_이미_대기_중이면_예외가_발생한다() {
-            // given
-            Time time = timeWithId(1L);
-            Theme theme = themeWithId(1L);
-            LocalDate date = LocalDate.now().plusYears(1);
-            ReservationCreateCommand command = new ReservationCreateCommand(
-                new ReserverName("예약자"), date, 1L, 1L);
-            when(timeRepository.findTimeByIdAndNotDeleted(1L)).thenReturn(Optional.of(time));
-            when(themeRepository.findThemeByIdAndNotDeleted(1L)).thenReturn(Optional.of(theme));
-            when(reservationRepository.existsReservationAndStatus(
-                any(Reservation.class), any(ReservationStatus.class)))
-                .thenAnswer(invocation -> {
-                    ReservationStatus status = invocation.getArgument(1);
-                    return status == ReservationStatus.WAITING;
-                });
-
-            // when & then
-            assertThatThrownBy(() -> reservationService.saveWaitingReservation(command))
-                .isInstanceOf(GeneralException.class)
-                .hasMessage("이미 대기 중인 이름, 날짜, 시간, 테마입니다.");
-        }
-
-        @Test
-        void 존재하지_않는_timeId로_대기_생성_시_파라미터_에러가_발생한다() {
-            // given
-            ReservationCreateCommand command = new ReservationCreateCommand(
-                new ReserverName("예약자"), LocalDate.now().plusDays(1), 999L, 999L);
-            when(timeRepository.findTimeByIdAndNotDeleted(999L)).thenReturn(Optional.empty());
-            when(themeRepository.findThemeByIdAndNotDeleted(999L)).thenReturn(Optional.empty());
-
-            // when & then
-            assertThatThrownBy(() -> reservationService.saveWaitingReservation(command))
-                .isInstanceOfSatisfying(GeneralParametersException.class, ex -> {
-                    assertThat(ex.getParameterErrors())
-                        .extracting(ParameterErrorResponseDto::parameter)
-                        .containsExactly("timeId", "themeId");
-                });
         }
     }
 
@@ -616,94 +548,6 @@ class ReservationServiceTest {
 
             // when & then
             assertThatThrownBy(() -> reservationService.cancelReservation(1L, new ReserverName("예약자")))
-                .isInstanceOf(GeneralException.class)
-                .hasMessage("지난 예약은 취소할 수 없습니다.");
-        }
-    }
-
-    @Nested
-    class 예약_대기_취소 {
-
-        @Test
-        void 예약_대기를_취소한다() {
-            // given
-            LocalDate futureDate = LocalDate.now().plusYears(1);
-            Time time = timeWithId(1L);
-            Theme theme = themeWithId(1L);
-            Reservation waiting = Reservation.reconstruct(
-                1L, new ReserverName("예약자"), futureDate, time, theme, ReservationStatus.WAITING);
-            Reservation canceledWaiting = Reservation.reconstruct(
-                1L, new ReserverName("예약자"), futureDate, time, theme, ReservationStatus.CANCELED);
-            when(reservationRepository.findReservationByIdAndNotDeleted(1L))
-                .thenReturn(Optional.of(waiting));
-            when(reservationRepository.update(any(Reservation.class))).thenReturn(canceledWaiting);
-
-            // when
-            ReservationCancelResponseDto result = reservationService.cancelWaitingReservation(1L, new ReserverName("예약자"));
-
-            // then
-            assertThat(result.id()).isEqualTo(1L);
-        }
-
-        @Test
-        void 존재하지_않는_예약_ID이면_예외가_발생한다() {
-            // given
-            when(reservationRepository.findReservationByIdAndNotDeleted(999L))
-                .thenReturn(Optional.empty());
-
-            // when & then
-            assertThatThrownBy(() -> reservationService.cancelWaitingReservation(999L, new ReserverName("예약자")))
-                .isInstanceOf(GeneralException.class)
-                .hasMessage("예약을 찾을 수 없습니다.");
-        }
-
-        @Test
-        void 예약자_이름이_다르면_예외가_발생한다() {
-            // given
-            LocalDate futureDate = LocalDate.now().plusDays(1);
-            Time time = timeWithId(1L);
-            Theme theme = themeWithId(1L);
-            Reservation waiting = Reservation.reconstruct(
-                1L, new ReserverName("예약자"), futureDate, time, theme, ReservationStatus.WAITING);
-            when(reservationRepository.findReservationByIdAndNotDeleted(1L))
-                .thenReturn(Optional.of(waiting));
-
-            // when & then
-            assertThatThrownBy(() -> reservationService.cancelWaitingReservation(1L, new ReserverName("다른사람")))
-                .isInstanceOf(GeneralException.class)
-                .hasMessage("예약을 취소할 권한이 없습니다.");
-        }
-
-        @Test
-        void 대기_예약이_아니면_예외가_발생한다() {
-            // given
-            LocalDate futureDate = LocalDate.now().plusDays(1);
-            Time time = timeWithId(1L);
-            Theme theme = themeWithId(1L);
-            Reservation active = Reservation.reconstruct(
-                1L, new ReserverName("예약자"), futureDate, time, theme, ReservationStatus.ACTIVE);
-            when(reservationRepository.findReservationByIdAndNotDeleted(1L))
-                .thenReturn(Optional.of(active));
-
-            // when & then
-            assertThatThrownBy(() -> reservationService.cancelWaitingReservation(1L, new ReserverName("예약자")))
-                .isInstanceOf(GeneralException.class)
-                .hasMessage("대기중인 예약이 아닙니다.");
-        }
-
-        @Test
-        void 지난_날짜의_대기_예약이면_예외가_발생한다() {
-            // given
-            LocalDate pastDate = LocalDate.now().minusDays(1);
-            Time time = timeWithId(1L);
-            Theme theme = themeWithId(1L);
-            Reservation pastWaiting = Reservation.reconstruct(
-                1L, new ReserverName("예약자"), pastDate, time, theme, ReservationStatus.WAITING);
-            when(reservationRepository.findReservationByIdAndNotDeleted(1L))
-                .thenReturn(Optional.of(pastWaiting));
-
-            // when & then
-            assertThatThrownBy(() -> reservationService.cancelWaitingReservation(1L, new ReserverName("예약자")))
                 .isInstanceOf(GeneralException.class)
                 .hasMessage("지난 예약은 취소할 수 없습니다.");
         }
